+++ src/components/fields/optimistic-markdown-input.js
@@ -1,49 +1,203 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
+import useDebouncedValue from 'Hooks/use-debounced-value';
+
 import MarkdownInput from '../inputs/markdown-input';
-import useOptimisticText from './use-optimistic-text';
 
-const OptimisticMarkdownInput = ({ value, onChange, ...props }) => {
-  const [state, setState] = React.useState({ isLoading: false, lastSavedResponse: value, inputState: value, error: null });
 
-  const onChangeWithSavedGoodResponse = (change) => {
-    setState({ isLoading: true });
-    if (onChange) {
-      return onChange(change).then(({status}) => {
-        if (status === 200) {
-          setState({ isLoading: false, lastSavedResponse: change, inputState: change })
-        } else {
-          setState({ ...state, isLoading: false })
-        }
-      }, (error) => {
-        setState({ ...state, isLoading: false })
-        throw error
-      });
-    }
-  }
+/*
+
+- takes in an initial state value 
+- takes in a way to update the server
+
+as users type (on change):
+- we ping the server with a trimmed version of the text
+- we display an untrimmed version
+
+if the server hits an error:
+- we display that error to the user
+- we continue to let the user type as normal
+
+on blur:
+- we revert the input back to the last successfully saved response
+- we hide the error
+
+*/
+
+const OptmisticMarkdownInput = ({ value, onChange, ...props }) => {
+  const {
+    inputValue, wrappedOnChange, optimisticError, wrappedOnBlur,
+  } = useOptimistValueOnChangeAndBlur({ value, asyncUpdate: onChange });
   
-  const onBlur = () => {
-    if (state.isLoading) {
-      onBlur();
-    } else {
-      setState({ ...state, inputState: state.lastSavedResponse, error: null })
-    }
-  }
-  
-  const [optimisticValue, optimisticOnChange, optimisticError] = useOptimisticText(state.inputState, onChangeWithSavedGoodResponse);
-  
-  React.useEffect(() => {
-    setState({ ...state, inputState: optimisticValue, error: optimisticError})
-  }, [optimisticValue, optimisticError])
-  console.log({ optimisticValue, optimisticError, ...state })
-  
-  return <MarkdownInput {...props} value={state.inputState} error={state.error} onChange={optimisticOnChange} onBlur={onBlur} />;
+  return <MarkdownInput {...props} value={inputValue} error={optimisticError} onChange={wrappedOnChange} optimisticError={optimisticError} onBlur={wrappedOnBlur} />;
 };
 
-OptimisticMarkdownInput.propTypes = {
+OptmisticMarkdownInput.propTypes = {
   value: PropTypes.string.isRequired,
   onChange: PropTypes.func.isRequired,
 };
+export default OptmisticMarkdownInput;
 
-export default OptimisticMarkdownInput;
+
+function useOptimistValueOnChangeAndBlur({ value, asyncUpdate }) {
+  const [nonAggressivelyTrimmedInputValue, onChangeWrappedWithTrimmedValue] = useNonAggressivelyTrimmedInputs(value, asyncUpdate);
+  const [inputValue, wrappedOnChange, wrappedOnBlur] = useRevertOnBlurWhenError(nonAggressivelyTrimmedInputValue, onChangeWrappedWithTrimmedValue);
+  const [optimisticValue, optimisticOnChange, optimisticError] = useOptimisticValueAndDebounceCallsToServer(inputValue, wrappedOnChange);
+
+  console.log(inputValue, optimisticValue, optimisticError)
+  return {
+    inputValue, wrappedOnChange, wrappedOnBlur, optimisticError,
+  };
+}
+
+function useNonAggressivelyTrimmedInputs(rawInput, asyncUpdate) {
+  const [untrimmedValue, setUntrimmedValue] = React.useState(rawInput);
+  
+  const displayedInputValue = rawInput === untrimmedValue.trim() ? untrimmedValue : rawInput;
+  const wrapAsyncUpdateWithTrimmedValue = (value) => {
+    setUntrimmedValue(value);
+    return asyncUpdate(value.trim());
+  };
+  return [displayedInputValue, wrapAsyncUpdateWithTrimmedValue];
+};
+
+function useRevertOnBlurWhenError(value, asyncUpdate, onBlur) {
+  const [state, setState] = React.useState({ 
+    status: null, 
+    lastSavedResponse: value, 
+    inputValue: value 
+  });
+
+  const wrappedOnChange = (change) => {
+    setState({ status: "loading" });
+    if (asyncUpdate) {
+      return asyncUpdate(change)
+        .then(response => {
+          console.log({ response })
+          if (response.status === 200) {
+            setState({ status: "loaded", lastSavedResponse: change, inputValue: change });
+          } else {
+            setState({ ...state, status: "error" });
+          }
+        }).catch(error => {
+          console.log("hit an error?", error)
+          setState({ ...state, status: "error" });
+          // throw error; // rethrow error so other funcs can use it later if necessary
+        });
+    }
+  }
+  
+  const wrappedOnBlur = () => {
+    // if we're still waiting for a response from the server, keep running this func
+    if (state.status === "loading") {
+      wrappedOnBlur();
+    } else {
+      setState({ ...state, inputValue: state.lastSavedResponse });
+      if (onBlur) { onBlur(); }
+    }
+  }
+  
+  React.useEffect(() => {
+    setState({ ...state, inputValue: value })
+  }, [value]);
+  
+  return [state.inputValue || value, wrappedOnChange, wrappedOnBlur];
+}
+
+
+/*
+  What this does:
+  - limit server calls to everytime state changes and it's been at least 500 ms
+  - but shows whatever the latest is as you type even if it hasn't saved yet
+  - returns an error message to be displayed to the user in case of error
+*/
+
+
+function useOptimisticValueAndDebounceCallsToServer(rawValueFromOnChange, setValueAsync) {
+  // store what is being typed in, along with an error message
+  // update undefined means that the field is unchanged from the 'real' value
+  const [state, setState] = React.useState({ updateToSave: undefined, error: null });
+  
+  // save a debounced version of what's being typed in
+  const debouncedUpdateToSave = useDebouncedValue(state.updateToSave, 500);
+  
+  const wrapOnChange = (change) => {
+    //save what's being typed in
+    setState({ ...state, updateToSave: change });
+    
+    const textHasUpdatedSinceSave = debouncedUpdateToSave !== undefined;
+    if (textHasUpdatedSinceSave) {
+      
+      // if the updateToSave changes during the async action then ignore the result
+      const setStateIfLastServerCallIsStillRelevant = (newState) => {
+        setState((prevState) => {
+          console.log(prevState, debouncedUpdateToSave, newState)
+          return prevState.updateToSave === debouncedUpdateToSave ? newState : prevState 
+        })
+      };
+      // this scope can't be async/await because it's an effect
+      setValueAsync(debouncedUpdateToSave).then(
+        (response) => {
+          console.log("saved!")
+          setStateIfLastServerCallIsStillRelevant({ updateToSave: undefined, error: null })
+          return response
+        },
+        (error) => {
+          console.log("hit an error")
+          const message = (error && error.response && error.response.data && error.response.data.message) || "Sorry! Something went wrong, try again later?"; 
+          console.log({ message })
+          setStateIfLastServerCallIsStillRelevant({ updateToSave: debouncedUpdateToSave, error: message });
+        },
+      );
+    }
+  }
+
+  const optimisticValue = state.updateToSave !== undefined ? state.updateToSave : rawValueFromOnChange;
+  
+
+  return [optimisticValue, wrapOnChange, state.error];
+};
+
+
+// function useOptimisticValueAndDebounceCallsToServerOLLLLLd(rawValueFromOnChange, setValueAsync) {
+//   // store what is being typed in, along with an error message
+//   // update undefined means that the field is unchanged from the 'real' value
+//   const [state, setState] = React.useState({ updateToSave: undefined, error: null });
+//   const debouncedUpdateToSave = useDebouncedValue(state.updateToSave, 500);
+  
+//   React.useEffect(() => {
+//     const textHasUpdatedSinceSave = debouncedUpdateToSave !== undefined;
+//     if (textHasUpdatedSinceSave) {
+//       // if the updateToSave changes during the async action then ignore the result
+//       const setStateIfLastServerCallIsStillRelevant = (newState) => {
+//         setState((prevState) => {
+//           console.log(prevState, debouncedUpdateToSave, newState)
+//           return prevState.updateToSave === debouncedUpdateToSave ? newState : prevState 
+//         })
+//       };
+//       // this scope can't be async/await because it's an effect
+//       setValueAsync(debouncedUpdateToSave).then(
+//         (response) => {
+//           console.log("saved!")
+//           setStateIfLastServerCallIsStillRelevant({ updateToSave: undefined, error: null })
+//           return response
+//         },
+//         (error) => {
+//           console.log("hit an error")
+//           const message = (error && error.response && error.response.data && error.response.data.message) || "Sorry! Something went wrong, try again later?"; 
+//           console.log({ message })
+//           setStateIfLastServerCallIsStillRelevant({ updateToSave: debouncedUpdateToSave, error: message });
+//         },
+//       );
+//     }
+//   }, [debouncedUpdateToSave]);
+
+//   const optimisticValue = state.updateToSave !== undefined ? state.updateToSave : rawValueFromOnChange;
+  
+//   const setOptimisticValue = (newValue) => {
+//     setState((prevState) => ({ ...prevState, updateToSave: newValue }));
+//   };
+//   return [optimisticValue, setOptimisticValue, state.error];
+// };
+
